+++ srp-reflection-support/src/main/java/org/srplib/reflection/deepcompare/comparators/ListDeepComparator.java
@@ -25,7 +25,10 @@ public class ListDeepComparator implements DeepComparator<List> {
     public void compare(List list1, List list2, DeepComparatorContext context) {
 
         if (list1.size() != list2.size()) {
-            context.registerMismatch(String.format("Lists have different size. Expected: %d actual: %d", list1.size(), list2.size()));
+            context.registerMismatch(
+                String.format("Lists have different size. Expected: %d actual: %d", list1.size(), list2.size()));
+
+            return;
         }
 
         for (int i = 0; i < list1.size(); i++) {